Split metal_backend tests + Qwen3.5 inference example for file-size cap
- crates/spark-runtime/src/metal_backend.rs: extracted the 3838-line
  `mod tests` body to a sibling directory. Tests are sharded by
  kernel family (parity_basic / quant / attention / attention_full /
  norms / vision / gdn) and integration tier (real_model_attention /
  vision / gemv / misc). Shared byte-conversion utilities + the
  synthetic MLX-int8 fixture builder live in `tests/helpers.rs`. All
  files in `tests/` are ≤500 LoC.
- crates/spark-runtime/examples/metal_qwen35_inference: split the
  monolithic `metal_qwen35_inference.rs` into a directory of sibling
  modules — `dims.rs` (compile-time dimensions), `full_attention.rs`
  (FA decoder layer + Scratch + LayerKvCache + forward), and
  `linear_attention.rs` (GDN decoder layer + state + LinScratch +
  forward). The Cargo example path was updated to
  `examples/metal_qwen35_inference/main.rs`.
- file-size-cap.yml: added two allow-list entries with rationale —
  `metal_backend.rs` (734 LoC, single irreducible `impl GpuBackend`)
  and `metal_qwen35_inference/main.rs` (700 LoC linear driver,
  splitting just inflates argument-list plumbing).

End-to-end inference still produces the MLX-matching output and the
24 metal_backend parity tests still pass.
