+++ .github/workflows/file-size-cap.yml
@@ -71,6 +71,29 @@ jobs:
             # extracting fixtures into a shared module that's only used
             # by this single test, adding indirection for no payoff.
             "crates/spark-server/tests/integration.rs"
+            # 734 LoC — `MetalGpuBackend` single-impl file. The body is
+            # a single `impl GpuBackend for MetalGpuBackend` which Rust
+            # forbids splitting across files; type aliases and the
+            # private inherent helpers (`current_cmd_buf`, `find_buffer`,
+            # `stream_index`) are tightly coupled to the trait methods'
+            # `Mutex`/`MTLBuffer`/`ObjCmdBuf` plumbing. The test module
+            # was already extracted to `metal_backend/tests/`; what's
+            # left is the irreducible backend surface. Tracked under
+            # the launch-day allow-list pending a free-function
+            # `ops/{alloc,copy,launch,events}.rs` extraction.
+            "crates/spark-runtime/src/metal_backend.rs"
+            # 700 LoC — end-to-end Qwen3.5 Metal inference example.
+            # The struct/forward-fn definitions for FA + GDN have
+            # already been extracted to siblings (`full_attention.rs`,
+            # `linear_attention.rs`, `dims.rs`); what's left is the
+            # `main` driver: env+config parsing, weight load, kernel
+            # handle resolution, prefill loop, decode loop, sampler,
+            # residual-stream dump scaffolding. Each phase is short and
+            # linear; splitting the driver just inflates argument-list
+            # plumbing for ~20 kernel handles + 4 scratch structs +
+            # several vec-of-layer aliases. Tracked under the
+            # launch-day allow-list.
+            "crates/spark-runtime/examples/metal_qwen35_inference/main.rs"
           )
 
           violations=0
